Trait frame_support_test::pallet::Config

pub trait Config: 'static + Eq + Clone {
    type BlockNumber: Parameter + Member + Default + MaybeSerializeDeserialize + MaxEncodedLen;
    type AccountId: Parameter + Member + MaxEncodedLen;
    type BaseCallFilter: Contains<Self::RuntimeCall>;
    type RuntimeOrigin: Into<Result<RawOrigin<Self::AccountId>, Self::RuntimeOrigin>> + From<RawOrigin<Self::AccountId>>;
    type RuntimeCall;
    type RuntimeEvent: Parameter + Member + IsType<<Self as Config>::RuntimeEvent> + From<Event<Self>>;
    type PalletInfo: PalletInfo;
    type DbWeight: Get<RuntimeDbWeight>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. The configuration trait.

Required Associated Types

type BlockNumber: Parameter + Member + Default + MaybeSerializeDeserialize + MaxEncodedLen

The block number type.

type AccountId: Parameter + Member + MaxEncodedLen

The account type.

type BaseCallFilter: Contains<Self::RuntimeCall>

The basic call filter to use in Origin.

type RuntimeOrigin: Into<Result<RawOrigin<Self::AccountId>, Self::RuntimeOrigin>> + From<RawOrigin<Self::AccountId>>

The runtime origin type.

type RuntimeCall

The runtime call type.

type RuntimeEvent: Parameter + Member + IsType<<Self as Config>::RuntimeEvent> + From<Event<Self>>

The runtime event type.

type PalletInfo: PalletInfo

The information about the pallet setup in the runtime.

type DbWeight: Get<RuntimeDbWeight>

The db weights.